solo_project/zigzag/requirements.txt
playwright
supabase
python-dotenv
pandas
beautifulsoup4
